+++ examples/dialogue_bot/src/dialogue/states/receive_location.rs
@@ -1,5 +1,5 @@
 use crate::dialogue::Dialogue;
-use teloxide::{macros::teloxide, prelude::*};
+use teloxide::prelude::*;
 
 #[derive(Generic)]
 pub struct ReceiveLocationState {