+++ examples/redis_remember_bot/src/transitions.rs
@@ -1,5 +1,4 @@
 use teloxide::prelude::*;
-use teloxide::macros::teloxide;
 
 use super::states::*;