+++ src/lib.rs
@@ -71,6 +71,7 @@ pub use teloxide_macros as macros;
 
 // FIXME(waffle): use `docsrs` here when issue with combine is resolved <https://github.com/teloxide/teloxide/pull/305#issuecomment-716172103>
 #[cfg_attr(all(teloxide_docsrs, feature = "nightly"), doc(cfg(feature = "macros")))]
+#[cfg(feature = "macros")]
 pub use teloxide_macros::teloxide;
 
 #[cfg(all(feature = "nightly", doctest))]